refactor: EmojiType 패키지 경로 변경 (record 밑에서 emoji 밑으로)
